mypy fix

+++ conmon/__main__.py
@@ -1016,7 +1016,7 @@ def monitor(args: List[str], replay=False) -> int:
     for item in ("conan_log", "report_json", "proc_json"):
         # copy replay file before opening or delete old ones
         path = replay_logfile(item, create_if_not_exists=replay)
-        if not replay and path.is_file():
+        if not replay and path and path.is_file():
             path.unlink()
     with filehandler("conan_log", hint="raw conan output") as fh:
         parser.process_streams(fh)
